estrutura do projeto finalizada
