+++ semana18/projeto-Labook/package.json
@@ -4,7 +4,7 @@
    "description": "",
    "main": "index.js",
    "scripts": {
-      "my-sql-setup":"tsc && node ./build/mySqlSetup.js",
+      "my-sql-setup": "tsc && node ./build/mySqlSetup.js",
       "start": "npm run build && node --inspect ./build/src/index.js",
       "dev": "clear && ts-node-dev ./src/index.ts",
       "build": "clear && echo \"Transpiling files...\" && tsc && echo \"Done!\" ",
