+++ semana18/projeto-Labook/src/routes.ts
@@ -1,0 +1,7 @@
+import { Router } from 'express'
+
+const routes = Router()
+
+// Rotas aqui !!!
+
+export default routes
